Update after testing

+++ src/scripts/views/pages/detail-post.js
@@ -1,3 +1,4 @@
+import Swal from 'sweetalert2';
 import UrlParser from '../../routes/url-parser';
 import MotorSource from "../../data/motor-source";
 import detailPostCreator from '../templates/detail-post-creator';
@@ -43,10 +44,30 @@ const DetailPost = {
     });
 
     deleteBtn.addEventListener('click', async () => {
-      try {
-        await MotorSource.deleteMotor(motor._id);
-      } catch (error) {
-        console.error('Error deleting motor:', error.message);
+      const isConfirmed = await Swal.fire({
+        icon: 'warning',
+        title: 'Are you sure?',
+        text: 'You are about to delete this post. This action cannot be undone.',
+        showCancelButton: true,
+        confirmButtonColor: '#d33',
+        cancelButtonColor: '#3085d6',
+        confirmButtonText: 'Yes, delete it!',
+        cancelButtonText: 'Cancel',
+      });
+    
+      // Jika pengguna mengonfirmasi penghapusan
+      if (isConfirmed.isConfirmed) {
+        try {
+          await MotorSource.deleteMotor(motor._id);
+        } catch (error) {
+          console.error('Error deleting motor:', error.message);
+          // Tambahkan penanganan kesalahan sesuai kebutuhan
+          Swal.fire({
+            icon: 'error',
+            title: 'Delete Failed',
+            text: 'Failed to delete the post. Please try again.',
+          });
+        }
       }
     });
 
@@ -54,12 +75,22 @@ const DetailPost = {
     if (motor.comments && motor.comments.length > 0) {
       deleteCommentBtn.addEventListener('click', async () => {
         try {
-          await MotorSource.deleteComment(motor._id, motor.comments.map(comment => comment._id));
-          // Jika berhasil menghapus komentar, reload halaman
-            
+          const confirmDeleteComment = await Swal.fire({
+            title: 'Are you sure?',
+            text: 'You wont be able to delete this!',
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#d33',
+            cancelButtonColor: '#3085d6',
+            confirmButtonText: 'Yes, delete it!',
+          });
+
+          if (confirmDeleteComment.isConfirmed) {
+            await MotorSource.deleteComment(motor._id, motor.comments.map(comment => comment._id));
+            this.afterRender();
+          }
         } catch (error) {
           console.error('Error deleting comment:', error.message);
-          // Tambahkan penanganan kesalahan sesuai kebutuhan, misalnya, menampilkan pesan kesalahan kepada pengguna
           alert('Failed to delete comment. Please try again.');
         }
       });
